requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
mysqlclient==2.2.1
python-slugify==8.0.1
gunicorn==21.2.0
dj-database-url==2.1.0
django-filter==23.5
PyMySQL==1.1.1
whitenoise==6.6.0
